engines/net: get rid of conversion warning on clang/OSX

+++ engines/net.c
@@ -485,7 +485,7 @@ static void store_udp_seq(struct netio_data *nd, struct io_u *io_u)
 	struct udp_seq *us;
 
 	us = io_u->xfer_buf + io_u->xfer_buflen - sizeof(*us);
-	us->magic = cpu_to_le64(FIO_UDP_SEQ_MAGIC);
+	us->magic = cpu_to_le64((uint64_t) FIO_UDP_SEQ_MAGIC);
 	us->bs = cpu_to_le64((uint64_t) io_u->xfer_buflen);
 	us->seq = cpu_to_le64(nd->udp_send_seq++);
 }
